+++ src/tike/ptycho/solvers/__init__.py
@@ -1,4 +1,4 @@
 """Contains different solver implementations."""
 
-from .cgrad import *
-from .lstsq_grad import *
+from .combined import *
+from .divided import *
